Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    int num1, num2, num3, avg;
+    float num1, num2, num3, avg;
     scanf("%d %d %d",&num1,&num2,&num3);
     avg=(num1+num2+num3)/3;
     printf("Average: %.2f\n", avg);
